+++ src/plastimatch/proj_image.cxx
@@ -100,7 +100,7 @@ raw_load (Proj_image *proj, const char* img_filename)
 
     /* Load pixels */
     rc = fread (proj->img, sizeof(float), proj->dim[0] * proj->dim[1], fp);
-    if (rc != proj->dim[0] * proj->dim[1]) {
+    if (rc != (size_t) (proj->dim[0] * proj->dim[1])) {
 	fprintf (stderr, "Couldn't load raster data for %s\n",
 		 img_filename);
 	exit (-1);
@@ -189,7 +189,7 @@ pfm_load (Proj_image *proj, const char* img_filename)
 
     /* Load pixels */
     rc = fread (proj->img, sizeof(float), proj->dim[0] * proj->dim[1], fp);
-    if (rc != proj->dim[0] * proj->dim[1]) {
+    if (rc != (size_t) (proj->dim[0] * proj->dim[1])) {
 	fprintf (stderr, "Couldn't load raster data for %s\n",
 		 img_filename);
 	exit (-1);
